Improved accuracy of sine calculation

+++ include/parsing.h
@@ -4,7 +4,7 @@
 #include "stdint.h"
 
 uint32_t parse_int(char *string, char c);
-float parse_float(char *string, char c);
+double parse_float(char *string, char c);
 void parse_string(char *parser, char *string, char c);
 
 #endif
